+++ jason/src/api/wasm/reconnect_handle.rs
@@ -45,8 +45,8 @@ impl ReconnectHandle {
     /// Tries to reconnect a [`RpcSession`] in a loop with a growing backoff
     /// delay.
     ///
-    /// The first attempt will be performed immediately, and second attempt will
-    /// be performed after `starting_delay_ms`.
+    /// The first attempt will be performed immediately, and the second attempt
+    /// will be performed after `starting_delay_ms`.
     ///
     /// Delay between reconnection attempts won't be greater than
     /// `max_delay_ms`.
@@ -54,8 +54,9 @@ impl ReconnectHandle {
     /// After each reconnection attempt, delay between reconnections will be
     /// multiplied by the given `multiplier` until it reaches `max_delay_ms`.
     ///
-    /// If `multiplier` is a negative number than `multiplier` will be
-    /// considered as `0.0`. This might cause busy loop so its not recommended.
+    ///
+    /// If `multiplier` is a negative number then it will be considered as
+    /// `0.0`. This might cause a busy loop, so it's not recommended.
     ///
     /// Max elapsed time can be limited with an optional `max_elapsed_time_ms`
     /// argument.